requirements.txt
google-generativeai
python-dotenv
pydantic
yfinance
requests
